Module grin_api::client

High level JSON/HTTP client API

Functions

create_post_request

create_post_request_ex

get
Helper function to easily issue a HTTP GET request against a given URL that returns a JSON object. Handles request building, JSON deserialization and response code checking. This function spawns a new Tokio runtime, which means it is pretty inefficient for multiple requests. In those situations you are probably better off creating a runtime once and spawning get_async tasks on it

get_async
Helper function to easily issue an async HTTP GET request against a given URL that returns a future. Handles request building, JSON deserialization and response code checking.

get_no_ret
Helper function to easily issue a HTTP GET request on a given URL that returns nothing. Handles request building and response code checking.

post
Helper function to easily issue a HTTP POST request with the provided JSON object as body on a given URL that returns a JSON object. Handles request building, JSON serialization and deserialization, and response code checking.

post_async
Helper function to easily issue an async HTTP POST request with the provided JSON object as body on a given URL that returns a future. Handles request building, JSON serialization and deserialization, and response code checking.

post_no_ret
Helper function to easily issue a HTTP POST request with the provided JSON object as body on a given URL that returns nothing. Handles request building, JSON serialization, and response code checking.

post_no_ret_async
Helper function to easily issue an async HTTP POST request with the provided JSON object as body on a given URL that returns a future. Handles request building, JSON serialization and deserialization, and response code checking.

send_request